# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FindingsResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FindingResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "review_id": {
            "description": "Opaque resource identifier. Clients must not infer meaning from it.",
            "format": "uuid",
            "title": "Review Id",
            "type": "string"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ReviewWarning"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "review_id",
          "items",
          "total",
          "warnings"
        ],
        "title": "FindingsResponse",
        "type": "object"
      },
...
      "ReviewCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "document_id": {
            "description": "Opaque resource identifier. Clients must not infer meaning from it.",
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "review_pack_id": {
            "description": "Opaque resource identifier. Clients must not infer meaning from it.",
            "format": "uuid",
            "title": "Review Pack Id",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "review_pack_id"
        ],
        "title": "ReviewCreateRequest",
        "type": "object"
      },
      "ReviewListItemResponse": {
        "additionalProperties": false,
        "description": "Строка истории проверок.\n\nИмя файла нужно, чтобы человек узнал свою проверку в списке: идентификатор\nдля этого бесполезен. Диагностики ошибки здесь нет — в списке достаточно\nстатуса, подробности видны на самой проверке.",
        "properties": {
          "document_filename": {
            "title": "Document Filename",
            "type": "string"
          },
          "document_id": {
            "description": "Opaque resource identifier. Clients must not infer meaning from it.",
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "findings_count": {
            "minimum": 0.0,
            "title": "Findings Count",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "description": "UTC timestamp in ISO 8601 format",
                "examples": [
                  "2026-09-03T12:00:00.000Z"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "queued_at": {
            "description": "UTC timestamp in ISO 8601 format",
            "examples": [
              "2026-09-03T12:00:00.000Z"
            ],
            "title": "Queued At",
            "type": "string"
          },
          "review_id": {
            "description": "Opaque resource identifier. Clients must not infer meaning from it.",
            "format": "uuid",
            "title": "Review Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "timed_out",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "review_id",
          "document_id",
          "document_filename",
          "status",
          "queued_at",
          "finished_at",
          "findings_count"
        ],
        "title": "ReviewListItemResponse",
        "type": "object"
      },
      "ReviewListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewListItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ReviewListResponse",
        "type": "object"
      },
...
      "ReviewResponse": {
        "additionalProperties": false,
        "properties": {
          "document_id": {
            "description": "Opaque resource identifier. Clients must not infer meaning from it.",
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewPublicError"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "description": "UTC timestamp in ISO 8601 format",
                "examples": [
                  "2026-09-03T12:00:00.000Z"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "poll_after_ms": {
            "anyOf": [
              {
                "minimum": 1000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poll After Ms"
          },
          "queued_at": {
            "description": "UTC timestamp in ISO 8601 format",
            "examples": [
              "2026-09-03T12:00:00.000Z"
            ],
            "title": "Queued At",
            "type": "string"
          },
          "review_id": {
            "description": "Opaque resource identifier. Clients must not infer meaning from it.",
            "format": "uuid",
            "title": "Review Id",
            "type": "string"
          },
          "review_pack_id": {
            "description": "Opaque resource identifier. Clients must not infer meaning from it.",
            "format": "uuid",
            "title": "Review Pack Id",
            "type": "string"
          },
          "stage": {
            "enum": [
              "waiting",
              "analysis",
              "result_ready",
              "finished"
            ],
            "title": "Stage",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "description": "UTC timestamp in ISO 8601 format",
                "examples": [
                  "2026-09-03T12:00:00.000Z"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "timed_out",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ReviewWarningResponse"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "review_id",
          "document_id",
          "review_pack_id",
          "status",
          "stage",
          "queued_at",
          "started_at",
          "finished_at"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DocReview API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/reviews": {
      "get": {
        "description": "Return recent reviews of the tenant, newest first.",
        "operationId": "list_reviews_api_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request payload is too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List Reviews",
        "tags": [
          "reviews"
        ]
      },
      "post": {
        "description": "Persist a queued job and return immediately; a worker runs it separately.",
        "operationId": "create_review_api_reviews_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request payload is too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Create Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews/{review_id}": {
      "get": {
        "description": "Return a small public lifecycle snapshot suitable for polling.",
        "operationId": "get_review_api_reviews__review_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "description": "Opaque resource identifier. Clients must not infer meaning from it.",
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request payload is too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Review",
        "tags": [
          "reviews"
        ]
      }
    },
...
    "/api/reviews/{review_id}/findings": {
      "get": {
        "description": "Return ordered public findings, or an empty list while analysis is pending.",
        "operationId": "get_review_findings_api_reviews__review_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "description": "Opaque resource identifier. Clients must not infer meaning from it.",
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request payload is too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Review Findings",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/reviews/{review_id}/retry": {
      "post": {
        "description": "Queue a new run linked to a terminal retriable failure.",
        "operationId": "retry_review_api_reviews__review_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "description": "Opaque resource identifier. Clients must not infer meaning from it.",
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request payload is too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Retry Review",
        "tags": [
          "reviews"
        ]
      }
    }
  }
}